[Fix #1410] Make registered cops aware of `AllCops: MigratedSchemaVersion`
When implementing #1383, the detection of `AllCops: MigratedSchemaVersion` was
initially considered only for cops related to migrations.

However, feedback received later, such as in #1410, indicated that it is also expected to apply to
`Style`, `Lint`, and other categories. This suggestion is reasonable, as warnings for migrated files
may not be limited to database columns but could also include Ruby programming logic.
Excluding `Style` and `Lint` from this consideration would not align with this feedback.

This PR modifies the behavior so that all registered cops can detect the value of `AllCops: MigratedSchemaVersion`.

Fixes #1410.
