+++ changelog/change_make_registered_cops_aware_of_all_cops_migrated_schema_version.md
@@ -1,0 +1,1 @@
+* [#1410](https://github.com/rubocop/rubocop-rails/issues/1410): Make registered cops aware of `AllCops: MigratedSchemaVersion`. ([@koic][])
