+++ lib/rubocop/cop/rails/dangerous_column_names.rb
@@ -14,7 +14,7 @@ module RuboCop
       #   # good
       #   add_column :users, :saved
       class DangerousColumnNames < Base # rubocop:disable Metrics/ClassLength
-        prepend MigrationsHelper
+        include MigrationsHelper
 
         COLUMN_TYPE_METHOD_NAMES = %i[
           bigint
